New bootloader protocol.
The main change is the use of a binary protocol, faster than the previous ASCII
protocol.
The I2C part has not been tested yet. Other features are planned and will be
added later.
The handling of several "contexts" has been partially added to blosh; it will be used to address several devices (using I2C) from a single blosh instance. Not tested. (This was implemented a long time ago and planned for the previous bootloader, but never committed.)

Only minor interface changes for blosh/robloc users:
 - features don't include supported commands anymore, they are now stored in a
   distinct value;
 - blosh option 'prog_check' (and associated robloc flag) has been removed:
   CRC check is always done, if supported by the device.